+++ src/api/controllers/shop.controller.ts
@@ -9,17 +9,22 @@ import {
   ParseIntPipe,
   Patch,
   Post,
+  Request,
+  UnauthorizedException,
   UseFilters,
+  UseGuards,
   UsePipes,
   ValidationPipe,
 } from '@nestjs/common';
-import { ApiBody, ApiResponse, ApiTags } from '@nestjs/swagger';
+import { ApiBearerAuth, ApiBody, ApiResponse, ApiTags } from '@nestjs/swagger';
+import { JwtAuthGuard } from 'auth/jwt-auth-guard';
 import { ApiErrorFilter } from 'exception/api-error-filter';
 import { ErrorStatus } from 'exception/error-status';
 import { ShopNotFoundError } from 'exceptions/shop-not-found';
 import { HttpErrorItem } from 'resources/http-error-item';
 import { Shop } from 'resources/shop';
 import { ShopService } from 'services/shop';
+import { RoleType } from 'types/roleType';
 import { CreateShopBody } from 'validators/shop/createShopBody';
 import { UpdateShopBody } from 'validators/shop/updateShopBody';
 
@@ -51,8 +56,19 @@ export class ShopController {
   @ApiResponse({ status: HttpStatus.CREATED, type: Shop })
   @ApiResponse({ status: HttpStatus.BAD_REQUEST, type: HttpErrorItem })
   @ApiResponse({ status: HttpStatus.INTERNAL_SERVER_ERROR, type: HttpErrorItem })
-  async post(@Body() shop: CreateShopBody): Promise<Shop> {
-    return Shop.from(await this.shopsService.createShop(shop));
+  @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth()
+  async post(
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-ignore
+    @Request() req,
+    @Body() shop: CreateShopBody,
+  ): Promise<Shop> {
+    if (req.user.role === RoleType.admin) {
+      return Shop.from(await this.shopsService.createShop(shop));
+    } else {
+      throw new UnauthorizedException();
+    }
   }
 
   @Patch('/:id')
@@ -62,8 +78,20 @@ export class ShopController {
   @ApiResponse({ status: HttpStatus.NOT_FOUND, type: HttpErrorItem })
   @ErrorStatus(ShopNotFoundError, HttpStatus.NOT_FOUND)
   @ApiBody({ type: UpdateShopBody })
-  async patch(@Param('id', ParseIntPipe) id: number, @Body() shop: UpdateShopBody): Promise<Shop> {
-    return Shop.from(await this.shopsService.updateShop(id, shop));
+  @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth()
+  async patch(
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-ignore
+    @Request() req,
+    @Param('id', ParseIntPipe) id: number,
+    @Body() shop: UpdateShopBody,
+  ): Promise<Shop> {
+    if (req.user.role === RoleType.admin || req.user.role === RoleType.vendor) {
+      return Shop.from(await this.shopsService.updateShop(id, shop));
+    } else {
+      throw new UnauthorizedException();
+    }
   }
 
   @Delete('/:id')
@@ -72,7 +100,18 @@ export class ShopController {
   @ApiResponse({ status: HttpStatus.INTERNAL_SERVER_ERROR, type: HttpErrorItem })
   @ErrorStatus(ShopNotFoundError, HttpStatus.NOT_FOUND)
   @HttpCode(HttpStatus.NO_CONTENT)
-  async delete(@Param('id', ParseIntPipe) id: number): Promise<void> {
-    await this.shopsService.deleteShop(id);
+  @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth()
+  async delete(
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-ignore
+    @Request() req,
+    @Param('id', ParseIntPipe) id: number,
+  ): Promise<void> {
+    if (req.user.role === RoleType.admin) {
+      await this.shopsService.deleteShop(id);
+    } else {
+      throw new UnauthorizedException();
+    }
   }
 }
